[ADD] Add ViewModel for Operations feature
This commit introduces a `ViewModel` for the Operations feature.

The `OperationViewModel` is created to manage the state of the `OperationsScreen`. It exposes an `OperationsState` which includes lists of `MenuItem` and `RepositoryItem` objects. The `ViewModel` also handles events such as selecting a repository.

The `lifecycle-viewmodel-compose` library has been added as a dependency to the `features/operations` module to support this `ViewModel`.

Minor UI adjustments have been made to the `OperationsScreen` to wrap the sticky header content with a `Surface` for better visual separation.