[EDIT] Add @Contextual annotation for AuthToken serialization
This commit adds the `@Contextual` annotation to the `accessToken` and `refreshToken` fields in `LoginResponseDto.kt`. This is necessary for proper serialization/deserialization of the `AuthToken` type when it's not a built-in serializable type.

Additionally, the `kotlin-serialization` plugin has been added to the `features/register/build.gradle.kts` file to enable serialization capabilities for this module.